Split expression function checking
