+++ docs/PHASE_PLAN.md
@@ -2210,6 +2210,10 @@ checked-in docs, tests, and commits only.
   `src/typechecker/expressions/method_call_support.rs`, keeping direct
   function-call checking separate from method, generic-method, and UFC
   resolution while preserving generic method fixture coverage.
+- Expression function checking now lives in
+  `src/typechecker/expressions/function_checking.rs`, keeping
+  `src/typechecker/expressions.rs` focused on expression dispatch while
+  preserving return/fallthrough and defer coverage.
 
 ## Current Phase
