+++ src/typechecker/expressions.rs
@@ -5,6 +5,7 @@ mod aggregate_support;
 mod call_support;
 mod call_validation;
 mod control_flow_support;
+mod function_checking;
 mod method_call_support;
 
 use crate::ast::expressions::StringPart;
@@ -17,90 +18,6 @@ use super::monomorphize_inference::InferenceConflict;
 use super::{BehaviorBound, TypeChecker};
 
 impl TypeChecker {
-    pub(crate) fn check_function(
-        &mut self,
-        name: &str,
-        params: &[Param],
-        return_type: &Option<AstType>,
-        body: &Expression,
-        _span: &Span,
-    ) -> Result<TypedFunction, Diagnostic> {
-        let ret_type = return_type
-            .as_ref()
-            .map(|t| self.resolve_type(t))
-            .unwrap_or(Type::Void);
-
-        self.current_return_type = Some(ret_type.clone());
-
-        // Push function scope with params
-        self.push_scope();
-        let mut typed_params = Vec::new();
-        for p in params {
-            let ty = self.resolve_type(&p.ty);
-            self.define_var(&p.name, ty.clone());
-            typed_params.push(TypedParam {
-                name: p.name.clone(),
-                ty,
-                span: p.span,
-            });
-        }
-
-        let typed_body = self.check_expr(body)?;
-        let body_block = match typed_body.kind {
-            TypedExprKind::Block(block) => block,
-            _ => TypedBlock {
-                ty: typed_body.ty.clone(),
-                span: typed_body.span,
-                statements: Vec::new(),
-                expr: Some(Box::new(typed_body)),
-            },
-        };
-
-        self.pop_scope();
-        self.current_return_type = None;
-
-        if ret_type != Type::Void && ret_type != Type::Never {
-            if let Some(expr) = &body_block.expr {
-                if expr.ty != Type::Never && !self.types_compatible(&ret_type, &expr.ty) {
-                    return Err(Diagnostic::error(
-                        "E3030",
-                        format!(
-                            "return type mismatch: expected `{}`, found `{}`",
-                            ret_type.display_name(),
-                            expr.ty.display_name()
-                        ),
-                        expr.span,
-                    ));
-                }
-            }
-
-            if !self.block_satisfies_return(&body_block, &ret_type) {
-                return Err(Diagnostic::error(
-                    "E3031",
-                    format!(
-                        "function `{}` must return `{}` on all non-error paths",
-                        name,
-                        ret_type.display_name()
-                    ),
-                    *_span,
-                ));
-            }
-        }
-
-        // Collect defers accumulated during this function's body (LIFO order)
-        let mut defers: Vec<TypedExpression> = self.pending_defers.drain(..).collect();
-        defers.reverse();
-
-        Ok(TypedFunction {
-            name: name.to_string(),
-            params: typed_params,
-            return_type: ret_type,
-            body: body_block,
-            defers,
-            span: *_span,
-        })
-    }
-
     pub(crate) fn check_expr(&mut self, expr: &Expression) -> Result<TypedExpression, Diagnostic> {
         match expr {
             Expression::IntLiteral { value, span } => Ok(TypedExpression {
